Unwrap strategies extraction
