+++ src/commonMain/kotlin/operations/numeric/Addition.kt
@@ -2,19 +2,10 @@ package operations.numeric
 
 import LogicOperation
 import asList
+import operations.numeric.unwrap.StrictUnwrapStrategy
 
-object Addition : LogicOperation, DoubleTypeSensitiveOperation {
+object Addition : LogicOperation, DoubleTypeSensitiveOperation, StrictUnwrapStrategy {
     override val key: String = "+"
 
     override fun invoke(expression: Any?, data: Any?): Any? = resultOrNull(expression.unwrapValues()) { it.sum() }
-
-    private fun Any?.unwrapValues() = asList.map(::unwrapValueFromList)
-
-    private fun unwrapValueFromList(value: Any?): Any? =
-        when (value) {
-            is Number -> value.toDouble()
-            is String -> value.toDoubleOrNull()
-            is List<*> -> unwrapValueFromList(value.firstOrNull())
-            else -> null
-        }
 }
