+++ src/commonMain/kotlin/operations/numeric/unwrap/StrictUnwrapStrategy.kt
@@ -1,0 +1,15 @@
+package operations.numeric.unwrap
+
+import asList
+
+interface StrictUnwrapStrategy: UnwrapStrategy {
+    override fun Any?.unwrapValues() = asList.map(::unwrap)
+
+    private fun unwrap(value: Any?): Any? =
+        when (value) {
+            is Number -> value.toDouble()
+            is String -> value.toDoubleOrNull()
+            is List<*> -> unwrap(value.firstOrNull())
+            else -> null
+        }
+}
