+++ src/commonMain/kotlin/operations/numeric/unwrap/UnwrapStrategy.kt
@@ -1,0 +1,5 @@
+package operations.numeric.unwrap
+
+interface UnwrapStrategy {
+    fun Any?.unwrapValues(): List<Any?>
+}
